app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
streamlit==1.39.0
plotly==5.24.1
python-dotenv==1.0.1
kafka-python==2.0.2
